fix: standardize all port types to u16
TCP ports are 0-65535, making u16 the correct type. Changed:
- DeploymentFixture.ports: Vec<i32> → Vec<u16>
- ServiceFixture.ports: Vec<(i32, i32)> → Vec<(u16, u16)>
- ServiceDef.port: Option<i32> → Option<u16>
- ServiceBuilder::port(), DeploymentFixture::port(): i32 → u16
- ServiceAssertion::has_port(): i32 → u16
- IngressTest::expect_backend(): i32 → u16

Uses i32::from(port) where k8s-openapi structs require i32 internally.

+++ src/context.rs
@@ -3712,7 +3712,7 @@ impl IngressTest {
     }
 
     /// Assert the ingress routes to the expected backend
-    pub async fn expect_backend(mut self, backend: &str, port: i32) -> Self {
+    pub async fn expect_backend(mut self, backend: &str, port: u16) -> Self {
         if self.err.is_some() {
             return self;
         }
@@ -3770,7 +3770,7 @@ impl IngressTest {
                                     let actual_port =
                                         service.port.and_then(|p| p.number).unwrap_or(0);
 
-                                    if actual_svc == expected_svc && actual_port == port {
+                                    if actual_svc == expected_svc && actual_port == i32::from(port) {
                                         found = true;
                                         break;
                                     }
